Refactoring the pytorch models distribution
