+++ src/simulai/models/_pytorch_models/__init__.py
@@ -1,0 +1,3 @@
+from ._deeponet import DeepONet, ResDeepONet, FlexibleDeepONet, ImprovedDeepONet
+from ._autoencoder import AutoencoderMLP, AutoencoderKoopman, AutoencoderVariational, AutoencoderCNN
+from ._miscellaneous import ImprovedDenseNetwork, ModelMaker, MetaModel, MoEPool
